# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: "Fluxa API"
  version: "1.0.0"
  description: |
    Fluxa is a multi-currency digital asset platform built on Stellar. It provides
    wallets, instant transfers, FX conversions, scheduled recurring payments,
    batch disbursements, fiat on/off-ramps, and a developer-friendly webhook system.

    All endpoints under `/v1` require a Bearer API key in the `Authorization` header.
  contact:
    name: "Fluxa Engineering"
    email: "engineering@fluxa.io"
    url: "https://github.com/fluxa/fluxa"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
servers:
  - url: http://localhost:3000
    description: "Local development"
  - url: https://api.fluxa.io
    description: "Production"
...
security:
  - BearerAuth: []
paths:
...
  /v1/wallets:
    post:
      operationId: createWallet
      summary: "Create a wallet"
      description: "Generates a new Stellar keypair, encrypts the secret, and returns the wallet."
      tags: ["Wallets"]
      responses:
        "201":
          description: "Wallet created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletCreated"
              example:
                id: "w_1a2b3c4d"
                public_key: "GBXCAFE7RJHBEVGT5ANBZ7YXDU7S3Y3VQWZ5J4YKJ2D6S7F8G9H0I"
                created_at: "2026-01-15T10:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/wallets/{id}:
    get:
      operationId: getWallet
      summary: "Get wallet details"
      description: "Returns the wallet metadata including public key and creation timestamp."
      tags: ["Wallets"]
      parameters:
        - $ref: "#/components/parameters/WalletID"
      responses:
        "200":
          description: "Wallet details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
              example:
                id: "w_1a2b3c4d"
                public_key: "GBXCAFE7RJHBEVGT5ANBZ7YXDU7S3Y3VQWZ5J4YKJ2D6S7F8G9H0I"
                created_at: "2026-01-15T10:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/wallets/{id}/balances:
    get:
      operationId: getWalletBalances
      summary: "Get wallet balances"
      description: "Returns all asset balances for the wallet from the Stellar network."
      tags: ["Wallets"]
      parameters:
        - $ref: "#/components/parameters/WalletID"
      responses:
        "200":
          description: "Wallet balances"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalances"
              example:
                wallet_id: "w_1a2b3c4d"
                balances:
                  - asset_code: "USDC"
                    issuer: "GA5ZSEJ6BHT2ZKGDABICHI62W25ZJUXUNFFYGLYTIEXP6CZMUWqmOvxM"
                    balance: "1250.0000000"
                  - asset_code: "XLM"
                    issuer: ""
                    balance: "42.5000000"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/wallets/{id}/trustlines:
    post:
      operationId: createTrustline
      summary: "Establish a trustline"
      description: "Adds a trustline for an asset on the wallet so it can hold that token."
      tags: ["Wallets"]
      parameters:
        - $ref: "#/components/parameters/WalletID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrustlineRequest"
            example:
              asset_code: "USDC"
              asset_issuer: "GA5ZSEJ6BHT2ZKGDABICHI62W25ZJUXUNFFYGLYTIEXP6CZMUWqmOvxM"
              limit: "100000"
      responses:
        "202":
          description: "Trustline submission accepted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrustlineResponse"
              example:
                wallet_id: "w_1a2b3c4d"
                asset_code: "USDC"
                asset_issuer: "GA5ZSEJ6BHT2ZKGDABICHI62W25ZJUXUNFFYGLYTIEXP6CZMUWqmOvxM"
                status: "submitted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/webhooks:
    post:
      operationId: registerWebhook
      summary: "Register a webhook endpoint"
      description: |
        Registers a new webhook endpoint. A signing secret is generated and
        returned — it will be used to sign webhook payloads via HMAC-SHA256.
      tags: ["Webhooks"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterWebhookRequest"
            example:
              url: "https://acme.com/hooks/fluxa"
              events:
                - "transfer.initiated"
                - "transfer.settled"
      responses:
        "201":
          description: "Webhook registered"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpointResponse"
              example:
                id: "wh_abc123"
                url: "https://acme.com/hooks/fluxa"
                secret: "whsec_abc123def456"
                events:
                  - "transfer.initiated"
                  - "transfer.settled"
                active: true
                created_at: "2026-01-15T10:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    get:
      operationId: listWebhooks
      summary: "List webhook endpoints"
      description: "Returns all registered webhook endpoints for the tenant."
      tags: ["Webhooks"]
      responses:
        "200":
          description: "List of webhook endpoints"
          content:
            application/json:
              schema:
                type: object
                properties:
                  endpoints:
                    type: array
                    items:
                      $ref: "#/components/schemas/WebhookEndpointSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/webhooks/{id}:
    delete:
      operationId: deleteWebhook
      summary: "Delete a webhook endpoint"
      description: "Permanently removes the webhook endpoint."
      tags: ["Webhooks"]
      parameters:
        - $ref: "#/components/parameters/WebhookID"
      responses:
        "204":
          description: "Webhook deleted"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
...
  securitySchemes:
    BearerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Pass a tenant API key as `Bearer <key>` in the `Authorization` header.
        Keys are prefixed with `sk_live_` (production) or `sk_test_` (sandbox).
    AdminAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Pass an admin API key as `Bearer <key>`. Admin keys are prefixed with
        `sk_admin_`.

  # ── Parameters ──────────────────────────────────────────────────────────
  parameters:
    WalletID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: "Wallet ID"
...
    WebhookID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: "Webhook endpoint ID"

  # ── Reusable Responses ──────────────────────────────────────────────────
  responses:
    BadRequest:
      description: "Bad request"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "BAD_REQUEST"
              message: "invalid request body"
    Unauthorized:
      description: "Unauthorized — missing or invalid API key"
      content:
        application/json:
          schema:
            type: object
          example:
            error: "missing or invalid authorization header"
    NotFound:
      description: "Resource not found"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "NOT_FOUND"
              message: "resource not found"
...
    InternalServerError:
      description: "Internal server error"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "INTERNAL_ERROR"
              message: "an unexpected error occurred"

  # ── Schemas ─────────────────────────────────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      required: ["error"]
      properties:
        error:
          type: object
          required: ["code", "message"]
          properties:
            code:
              type: string
              enum:
                - "BAD_REQUEST"
                - "NOT_FOUND"
                - "UNAUTHORIZED"
                - "FORBIDDEN"
                - "INTERNAL_ERROR"
                - "QUOTE_EXPIRED"
                - "QUOTE_ALREADY_USED"
              description: "Machine-readable error code"
            message:
              type: string
              description: "Human-readable error message"

    # ── Health ────────────────────────────────────────────────────────────
...
    Wallet:
      type: object
      properties:
        id:
          type: string
        public_key:
          type: string
          description: "Stellar public key"
        created_at:
          type: string
          format: date-time
    WalletCreated:
      type: object
      properties:
        id:
          type: string
        public_key:
          type: string
        created_at:
          type: string
          format: date-time
    WalletBalances:
      type: object
      properties:
        wallet_id:
          type: string
        balances:
          type: array
          items:
            $ref: "#/components/schemas/Balance"
    Balance:
      type: object
      properties:
        asset_code:
          type: string
          description: "Asset code (e.g. USDC, XLM)"
        issuer:
          type: string
          description: "Issuer public key (empty for native XLM)"
        balance:
          type: string
          description: "Available balance as decimal string"
    CreateTrustlineRequest:
      type: object
      required: ["asset_code", "asset_issuer"]
      properties:
        asset_code:
          type: string
        asset_issuer:
          type: string
          description: "Issuer public key"
        limit:
          type: string
          description: "Trust line limit (optional)"
    TrustlineResponse:
      type: object
      properties:
        wallet_id:
          type: string
        asset_code:
          type: string
        asset_issuer:
          type: string
        status:
          type: string
          enum: ["submitted"]

    # ── Fiat ──────────────────────────────────────────────────────────────
...
    RegisterWebhookRequest:
      type: object
      required: ["url"]
      properties:
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum:
              - "transfer.initiated"
              - "transfer.settled"
              - "transfer.failed"
              - "wallet.funded"
              - "conversion.completed"
          description: |
            Event types to subscribe to. Omit or pass an empty array to
            receive all events.
    WebhookEndpointResponse:
      type: object
      description: "Returned on create — includes the signing secret"
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        secret:
          type: string
          description: "HMAC-SHA256 signing secret (shown once)"
        events:
          type: array
          items:
            type: string
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
    WebhookEndpointSummary:
      type: object
      description: "Returned on list — secret is never included"
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
...